+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+### Added
+- `Deserializer` can now be configured to allow (or disallow) zero-copy deserialization.
 
 ## 0.4.0 - 2023-04-06
 ### Added